+++ lib/source/hmac_prng.c
@@ -39,43 +39,43 @@
  * min bytes in the seed string.
  * MIN_SLEN*8 must be at least the expected security level.
  */
-static const unsigned int MIN_SLEN = 32;
+static const uint32_t MIN_SLEN = 32;
 
 /*
  * max bytes in the seed string;
  * SP800-90A specifies a maximum of 2^35 bits (i.e., 2^32 bytes).
  */
-static const unsigned int MAX_SLEN = UINT32_MAX;
+static const uint32_t MAX_SLEN = UINT32_MAX;
 
 /*
  * max bytes in the personalization string;
  * SP800-90A specifies a maximum of 2^35 bits (i.e., 2^32 bytes).
  */
-static const unsigned int MAX_PLEN = UINT32_MAX;
+static const uint32_t MAX_PLEN = UINT32_MAX;
 
 /*
  * max bytes in the additional_info string;
  * SP800-90A specifies a maximum of 2^35 bits (i.e., 2^32 bytes).
  */
-static const unsigned int MAX_ALEN = UINT32_MAX;
+static const uint32_t MAX_ALEN = UINT32_MAX;
 
 /*
  * max number of generates between re-seeds;
  * TinyCrypt accepts up to (2^32 - 1) which is the maximal value of
- * a 32-bit unsigned int variable, while SP800-90A specifies a maximum of 2^48.
+ * a 32-bit uint32_t variable, while SP800-90A specifies a maximum of 2^48.
  */
-static const unsigned int MAX_GENS = UINT32_MAX;
+static const uint32_t MAX_GENS = UINT32_MAX;
 
 /*
  * maximum bytes per generate call;
  * SP800-90A specifies a maximum up to 2^19.
  */
-static const unsigned int  MAX_OUT = (1 << 19);
+static const uint32_t  MAX_OUT = ((uint32_t)1 << 19);
 
 /*
  * Assumes: prng != NULL
  */
-static void update(TCHmacPrng_t prng, const uint_least8_t *data, unsigned int datalen, const uint_least8_t *additional_data, unsigned int additional_datalen)
+static void update(TCHmacPrng_t prng, const uint_least8_t *data, uint32_t datalen, const uint_least8_t *additional_data, uint32_t additional_datalen)
 {
 	const uint_least8_t separator0 = 0x00;
 	const uint_least8_t separator1 = 0x01;
@@ -129,7 +129,7 @@ static void update(TCHmacPrng_t prng, const uint_least8_t *data, unsigned int da
 
 int tc_hmac_prng_init(TCHmacPrng_t prng,
 		      const uint_least8_t *personalization,
-		      unsigned int plen)
+		      uint32_t plen)
 {
 
 	/* input sanity check: */
@@ -153,9 +153,9 @@ int tc_hmac_prng_init(TCHmacPrng_t prng,
 
 int tc_hmac_prng_reseed(TCHmacPrng_t prng,
 			const uint_least8_t *seed,
-			unsigned int seedlen,
+			uint32_t seedlen,
 			const uint_least8_t *additional_input,
-			unsigned int additionallen)
+			uint32_t additionallen)
 {
 
 	/* input sanity check: */
@@ -189,9 +189,9 @@ int tc_hmac_prng_reseed(TCHmacPrng_t prng,
 	return TC_CRYPTO_SUCCESS;
 }
 
-int tc_hmac_prng_generate(uint_least8_t *out, unsigned int outlen, TCHmacPrng_t prng)
+int tc_hmac_prng_generate(uint_least8_t *out, uint32_t outlen, TCHmacPrng_t prng)
 {
-	unsigned int bufferlen;
+	uint32_t bufferlen;
 
 	/* input sanity check: */
 	if (out == (uint_least8_t *) 0 ||
